ServerApp/requirements/local.txt
coverage==5.5 
django-debug-toolbar==3.2